refactor: extract detachReceipt and replaceReceipt into IOU/Receipt.ts
Break out receipt-related functions from the monolithic IOU/index.ts
into a dedicated Receipt.ts module as part of the ongoing effort to
reduce the file size (#72804).
